# db: use pointer-receiver for range key sorts