Define the sort.Interface on a pointer receiver when sorting range keys. This
prevents an allocation. Benchmark results from running the benchmark in #1824:

```
name                                                  old time/op    new time/op    delta
Iterator_RangeKeyMasking/range-keys-suffixes=@10-10     2.52ms ± 1%    2.33ms ± 2%   -7.60%  (p=0.000 n=10+10)
Iterator_RangeKeyMasking/range-keys-suffixes=@50-10     2.36ms ± 1%    2.24ms ± 1%   -4.86%  (p=0.000 n=10+9)
Iterator_RangeKeyMasking/range-keys-suffixes=@75-10     2.23ms ± 2%    2.23ms ± 3%     ~     (p=0.905 n=9+10)
Iterator_RangeKeyMasking/range-keys-suffixes=@100-10    2.11ms ± 3%    2.10ms ± 1%     ~     (p=0.780 n=10+9)

name                                                  old alloc/op   new alloc/op   delta
Iterator_RangeKeyMasking/range-keys-suffixes=@10-10      372kB ± 0%     122kB ± 2%  -67.32%  (p=0.000 n=8+9)
Iterator_RangeKeyMasking/range-keys-suffixes=@50-10      257kB ± 1%     117kB ± 1%  -54.38%  (p=0.000 n=10+10)
Iterator_RangeKeyMasking/range-keys-suffixes=@75-10      182kB ± 2%     116kB ± 3%  -36.16%  (p=0.000 n=10+10)
Iterator_RangeKeyMasking/range-keys-suffixes=@100-10     110kB ± 1%     110kB ± 1%     ~     (p=0.190 n=10+10)

name                                                  old allocs/op  new allocs/op  delta
Iterator_RangeKeyMasking/range-keys-suffixes=@10-10      7.63k ± 0%     0.07k ± 0%  -99.10%  (p=0.000 n=9+9)
Iterator_RangeKeyMasking/range-keys-suffixes=@50-10      4.26k ± 0%     0.07k ± 0%  -98.40%  (p=0.000 n=10+8)
Iterator_RangeKeyMasking/range-keys-suffixes=@75-10      2.12k ± 0%     0.07k ± 2%  -96.78%  (p=0.000 n=8+10)
Iterator_RangeKeyMasking/range-keys-suffixes=@100-10      72.0 ± 0%      67.0 ± 0%   -6.94%  (p=0.000 n=10+10)
```